requirements.txt
torch==2.2.1
torchvision
tensorboard
transformers==4.36.0
sentencepiece==0.2.0
tokenizers>=0.12.1
huggingface-hub==0.20.3
accelerate==0.27.2
datasets==2.17.1
bitsandbytes==0.43.1
fast-hadamard-transform==1.0.4.post1
fschat==0.2.34
sacrebleu>=1.5.0
evaluate>=0.4.0
scikit-learn>=0.24.1
sqlitedict
peft>=0.2.0
jsonlines
numexpr
pybind11>=2.6.2
pytablewriter
rouge-score>=0.0.4
tqdm-multiprocess
zstandard
texttable
toml
attributedict
protobuf
auto-gptq @ git+https://github.com/PanQiWei/AutoGPTQ.git@v0.6.0
matplotlib
img2pdf